buf.h now uses size_t for array sizes.

+++ buf.c
@@ -1,8 +1,8 @@
 #include "buf.h"
 
-void buf_do_realloc_(void **a, int nr, int sz)
+void buf_do_realloc_(void **a, size_t nr, size_t sz)
 {
-	int *p = realloc(*a ? buf_raw_(*a) : NULL, 2 * sizeof(int) + nr * sz);
+	size_t *p = realloc(*a ? buf_raw_(*a) : NULL, 2 * sizeof(size_t) + nr * sz);
 	p[0] = nr;
 	if (!*a)
 		p[1] = 0;
